break up bigish component, quick code review changes
